Separated the functionalities of the shader utilities into two (factory/module). Also removed the gem::particles namespace for reusability.
