+++ ParticleSystem.vcxproj
@@ -32,7 +32,8 @@
     <ClCompile Include="src\rain_emitter.cc" />
     <ClCompile Include="src\random_fountain_emitter.cc" />
     <ClCompile Include="src\scene.cc" />
-    <ClCompile Include="src\shader.cc" />
+    <ClCompile Include="src\shader_factory.cc" />
+    <ClCompile Include="src\shader_module.cc" />
     <ClCompile Include="src\spherical_stream_emitter.cc" />
     <ClCompile Include="src\core_opengl_renderer.cc" />
   </ItemGroup>
@@ -56,9 +57,10 @@
     <ClInclude Include="src\random_fountain_emitter.hh" />
     <ClInclude Include="src\renderer.hh" />
     <ClInclude Include="src\scene.hh" />
-    <ClInclude Include="src\shader.hh" />
+    <ClInclude Include="src\shader_factory.hh" />
     <ClInclude Include="src\emitter.hh" />
     <ClInclude Include="src\particle_pool.hh" />
+    <ClInclude Include="src\shader_module.hh" />
     <ClInclude Include="src\singleton.hh" />
     <ClInclude Include="src\spherical_stream_emitter.hh" />
     <ClInclude Include="src\core_opengl_renderer.hh" />
